correcao transferencia de chassi
correcao transferencia de chassi

+++ Renave.Anfir/Models/EnvioAutorizacao.cs
@@ -12,6 +12,6 @@ namespace Renave.Anfir.Models
         public string cpfOperadorResponsavel { get; set; }
         public int idEstoqueOrigem { get; set; }
         public bool paraCancelamentoDeTransferencia { get; set; }
-        public double? valorProduto { get; set; }
+        public decimal? valorProduto { get; set; }
     }
 }
